requirements.txt
torch==2.0.1
torch_geometric==2.5.2
torch_cluster==1.6.3
torch-scatter==2.1.1+pt20cu117
torch-sparse==0.6.17+pt20cu117
torch-spline-conv==1.2.2+pt20cu117
pandas==1.3.5
numpy==1.24.4 
scikit_learn==1.3.0
PyGCL==0.1.2


